--- src/components/SprintSelect/index.tsx
@@ -1,135 +1,0 @@
-import { Select, SimpleGrid, VStack } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
-import { setupAPIMetrics } from "../../services/api";
-import { tokenService } from "../../services/auth/tokenService";
-
-export interface WorkRelations {
-  rel: string;
-  source: string;
-  target: {
-    id: number;
-    url: string;
-  };
-}
-interface Task {
-  ID: string;
-  Title: string;
-  "Work Item Type": string;
-  State: string;
-  "State Change Date": string;
-  Area: string;
-  "Iteration Path": string;
-  "Activated By": string;
-  "Activated Date": string;
-  "Assigned To": string | undefined;
-  "Changed By": string;
-  "Changed Date": string;
-  "Completed Work": string | undefined;
-  "Created By": string;
-  "Created Date": string;
-  Description: string | undefined;
-  Reason: string;
-  "Story Points": number | undefined | string;
-  "Cycle Time": number | undefined;
-  "Sprint Start Date": string;
-  Tags: string;
-  Activity: string;
-}
-
-interface Team {
-  id: string;
-}
-
-interface SelectSprintProps {
-  teamId: Team;
-  setworkItemRelations?: (itemRelations: WorkRelations[]) => void;
-  setTasks?: (task: Task[]) => void;
-}
-
-export interface Iterations {
-  id: string;
-  name: string;
-  path: string;
-  attributes: {
-    startDate: string;
-    finishDate: string;
-    timeFrame: string;
-  };
-  url: string;
-}
-
-const token = tokenService.getToken()
-const project_id = tokenService.getProjectId()
-const organization = tokenService.getOrganization()
-
-const axiosInstance = setupAPIMetrics({organization, project_id ,token} );
-
-export default function SprintSelect({
-  teamId,
-  setworkItemRelations,
-  setTasks,
-}: SelectSprintProps) {
-  const [selectedSprint, setSeletedSprint] = useState("");
-  const [sprint, setSprint] = useState<Iterations[]>([]);
-
-  useEffect(() => {
-    axiosInstance
-      .get(
-        `https://dev.azure.com/${organization}/${project_id}/${teamId}/_apis/work/teamsettings/iterations?api-version=6.0`
-      )
-      .then((response) => {
-        if(response.status === 200) {
-          setSprint(response.data.value.reverse());
-        } 
-      }).catch(error => {
-        console.warn(error.response)
-    });
-  }, [teamId]);
-
-  const handleChange = (event: any) => {
-    setSeletedSprint(event.target.value);
-
-    axiosInstance
-      .get(
-        `https://dev.azure.com/${organization}/${project_id}/${teamId}/_apis/work/teamsettings/iterations/${event.target.value}/workitems?api-version=6.0-preview.1`
-      )
-      .then((response) => {
-        if(response.status === 200) {
-          setworkItemRelations(response.data.workItemRelations)
-        }
-      }).catch(error => {
-        console.warn(error.response)
-    });
-
-    setTasks([]);
-  };
-
-  return (
-    <>
-      <VStack spacing="8">
-        <SimpleGrid
-          minChildWidth="240px"
-          spacing={["6", "8"]}
-          alignSelf="flex-start"
-        >
-          <VStack spacing={3}>
-            <Select
-              placeholder="Sprints"
-              size="lg"
-              onChange={(ev) => handleChange(ev)}
-              value={selectedSprint}
-            >
-              {sprint.map((item: Iterations) => {
-                return (
-                  <option key={item.id} value={item.id}>
-                    {item.name}
-                  </option>
-                );
-              })}
-            </Select>
-          </VStack>
-        </SimpleGrid>
-      </VStack>
-    </>
-  );
-}
