+++ src/pages/sprintReport.tsx
@@ -1,25 +1,16 @@
 import { withSession } from "../services/auth/session";
 import {
   Box,
-  Divider,
   Flex,
-  Tab,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Tabs,
-  Text,
 } from "@chakra-ui/react";
 
 import { Header } from "../components/Header";
 import { Grid, GridItem } from "@chakra-ui/react";
 import { useState } from "react";
-import SelectSprint from "../components/SelectSprint";
-import { GraphicBoard } from "../components/GraphicBoard";
-import { UsHistory } from "../components/UsHistory";
-import Report from "../data/report";
-import { Revisions } from "../components/Revisions";
 import { TeamsProvider } from "../contexts/TeamsContext";
+import ReportTabs from "../components/ReportTabs";
+import SelectSprintForm from "../components/SelectSprintForm";
+import TeamSelect from "../components/TeamSelect";
 
 interface Task {
   ID: string;
@@ -46,10 +37,32 @@ interface Task {
   Activity: string;
 }
 
+interface Team {
+  description: string;
+  id: string;
+  identityUrl: string;
+  name: string;
+  projectId: string;
+  projectName: string;
+  url: string;
+}
+
+export interface Iterations {
+  id: string;
+  name: string;
+  path: string;
+  attributes: {
+    startDate: string;
+    finishDate: string;
+    timeFrame: string;
+  };
+  url: string;
+}
+
 export default function SprintReport() {
   const [tasks, setTasks] = useState<Task[]>([]);
-  const [sprintTeam, setSprintTeam] = useState("");
-  const report = new Report();
+  const [sprintTeam, setSprintTeam] = useState<Team>();
+  const [sprint, setSprint] = useState<Iterations[]>([]);
 
   return (
     <TeamsProvider>
@@ -59,86 +72,13 @@ export default function SprintReport() {
             <Header />
           </GridItem>
           <GridItem colSpan={5} >
-            <>
-              <Flex direction="column"  justify="center" gap="2">
-                <SelectSprint
-                  setTasks={setTasks}
-                  setSprintTeam={setSprintTeam}
-                />
-                {sprintTeam && (
-                  <Tabs size="md" variant="enclosed" bg="white">
-                    <TabList>
-                      <Tab>General</Tab>
-                      <Tab>State Graph</Tab>
-                      <Tab>Board Column</Tab>
-                      <Tab>Bugs</Tab>
-                      <Tab>Improvements</Tab>
-                      <Tab>Not Expected</Tab>
-                      {/* <Tab>Related Bug</Tab> */}
-                      <Tab>Bugs - Roots Causes</Tab>
-                    </TabList>
-                    <TabPanels>
-                      <TabPanel>
-                        <GraphicBoard tasks={tasks} />
-                      </TabPanel>
-                      <TabPanel>
-                        <UsHistory tasks={tasks} workItemType="User Story" />
-                      </TabPanel>
-                      <TabPanel>
-                        <Revisions tasks={tasks} workItemType="User Story" />
-                      </TabPanel>
-                      <TabPanel>
-                        <UsHistory tasks={tasks} workItemType="Bug" />
-                      </TabPanel>
-                      <TabPanel>
-                        {
-                          report.returnAllTasksByWorkItemTag(tasks, "Melhoria").map((item, key) => {
-                            return (
-                              <Box key={key} mb="8">
-                                <Text>{item.ID} - {item.Title}</Text>
-                                <Divider />
-                              </Box>
-                            )
-                          })
-                        }
-                      </TabPanel>
-                      <TabPanel>
-                        {
-                          report.returnAllTasksByWorkItemTag(tasks, "não previsto").map((item, key) => {
-                            return (
-                              <Box key={key} mb="8">
-                                <Text>{item.ID} - {item.Title}</Text>
-                                <Divider />
-                              </Box>
-                            )
-                          })
-                        }
-                      </TabPanel>
-                      {/* <TabPanel>
-                      {report
-                        .returnAllTasksByWorkItemType(tasks, "User Story")
-                        .map((task, key) => {
-                          return <RelatedBugs key={key} task={task} />;
-                        })}
-                    </TabPanel> */}
-                      <TabPanel>
-                        {report
-                          .returnBugs(tasks)
-                          .map((bug, key) => {
-                            return (
-                              <Box key={key} mb="8">
-                                <Text>{bug.ID} - {bug.Title}</Text>
-                                <Text><strong>Root Cause: </strong>{bug.Activity}</Text>
-                                <Divider />
-                              </Box>
-                            )
-                          })}
-                      </TabPanel>
-                    </TabPanels>
-                  </Tabs>
-                )}
-              </Flex>
-            </>
+            <Flex direction="column" justify="center" gap="2">
+              <Box display="flex" mt="1px" bg="white" p={["6", "8"]} gap="5">
+                <TeamSelect setSprint={setSprint} setTask={setTasks} setTeam={setSprintTeam} />
+                {sprintTeam && <SelectSprintForm  teamId={sprintTeam} sprint={sprint} setTasks={setTasks} />}
+              </Box>
+              {tasks.length !== 0 && <ReportTabs tasks={tasks} />}
+            </Flex>
           </GridItem>
         </Grid>
       </Flex>
